refactor and add reusable close button
