+++ website/components/reusable/InputField.vue
@@ -9,7 +9,6 @@
 </template>
 
 <script>
-
 export default {
   name: 'InputField',
   components: {
